moved theme, styled icon and typography
