+++ components/anchor/demo/onClick.md
@@ -19,14 +19,14 @@ import { Anchor } from 'antd';
 const { Link } = Anchor;
 
 const handleClick = (
-  e: React.MouseEvent<HTMLElement>,
-  link: {
-    title: React.ReactNode;
-    href: string;
-  },
-) => {
-  e.preventDefault();
-  console.log(link);
+    e: React.MouseEvent<HTMLElement>,
+    link: {
+      title: React.ReactNode;
+      href: string;
+    },
+  ) => {
+    e.preventDefault();
+    console.log(link);
 };
 
 ReactDOM.render(
