+++ components/auto-complete/__tests__/focus.test.js
@@ -24,7 +24,10 @@ describe('AutoComplete children could be focus', () => {
   it('focus() and onFocus', () => {
     const handleFocus = jest.fn();
     const wrapper = mount(<AutoComplete onFocus={handleFocus} />, { attachTo: container });
-    wrapper.find('input').instance().focus();
+    wrapper
+      .find('input')
+      .instance()
+      .focus();
     jest.runAllTimers();
     expect(handleFocus).toHaveBeenCalled();
   });
@@ -32,9 +35,15 @@ describe('AutoComplete children could be focus', () => {
   it('blur() and onBlur', () => {
     const handleBlur = jest.fn();
     const wrapper = mount(<AutoComplete onBlur={handleBlur} />, { attachTo: container });
-    wrapper.find('input').instance().focus();
+    wrapper
+      .find('input')
+      .instance()
+      .focus();
     jest.runAllTimers();
-    wrapper.find('input').instance().blur();
+    wrapper
+      .find('input')
+      .instance()
+      .blur();
     jest.runAllTimers();
     expect(handleBlur).toHaveBeenCalled();
   });
