+++ components/footer/__tests__/index.test.js
@@ -5,7 +5,7 @@ import rtlTest from '../../../tests/shared/rtlTest';
 
 import { Footer } from '..';
 import { DefaultAppLayoutProvider } from '../../layout';
-import { ThemeProvider } from '../../style/themes/varnish';
+import { ThemeProvider } from '../../varnish';
 
 describe('Footer', () => {
   mountTest(Footer);
