+++ components/varnish/shape.ts
