+++ site/theme/template/varnish/BreakpointDisplay.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import styled from 'styled-components';
-import { Spacing } from '../../../../components/style/themes/varnish/spacing';
-import { breakpoints } from '../../../../components/style/themes/varnish/breakpoints';
+import { Spacing } from '../../../../components/varnish/spacing';
+import { breakpoints } from '../../../../components/varnish/breakpoints';
 
 // eslint-disable-next-line import/prefer-default-export
 export class BreakpointDisplay extends React.PureComponent {
