+++ tests/shared/rtlTest.tsx
@@ -3,7 +3,7 @@ import Moment from 'moment';
 import MockDate from 'mockdate';
 import { mount } from 'enzyme';
 import ConfigProvider from '../../components/config-provider';
-import { ThemeProvider } from '../../components/style/themes/varnish';
+import { ThemeProvider } from '../../components/varnish';
 
 // eslint-disable-next-line jest/no-export
 export default function rtlTest(Component: React.ComponentType, mockDate?: boolean) {
